把生成颜色和位置的代码移出去。
